requirements.txt
aniso8601==9.0.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dataclasses==0.8
Django==1.11
Faker==8.2.1
Flask==2.0.1
Flask-MySQL==1.5.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.26
protobuf==3.17.3
pycparser==2.20
PyMySQL==1.0.2
pyodbc==4.0.31
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.22
text-unidecode==1.3
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
wincertstore==0.2
zipp==3.5.0
